infra: move app/search.py to infra/search_index.py
The Whoosh full-text index client is an external client against an on-disk
index, i.e. Level 5 infra per the target-architecture doc — not a service.
It cannot become services/search.py because five crud modules import
do_search from it and crud must not import services (ratchet-enforced).

Pure relocation: schemas, do_search and the index path are unchanged.
services/search.py keeps its existing role (index refresh / write side).

Refs docs/proposals/2026-07-15-target-architecture.md step 4 item 5.
